Basic GUI Implementation completed!
